Integration Test

+++ src/main/java/com/intelligent/ecommerce/dto/order/response/OrderReportResponse.java
@@ -1,5 +1,6 @@
 package com.intelligent.ecommerce.dto.order.response;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 import com.intelligent.ecommerce.enums.OrderStatus;
@@ -14,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class OrderReportResponse {
     private Long id;
     private Long customerId;
-    private Double totalAmount;
+    private BigDecimal totalAmount;
     private OrderStatus status;
     private Instant createdAt;
 }
